完善代码 0.补充训练模型代码
by organwalk 2023-10-09

+++ train_code/utils/data_utils.py
@@ -15,11 +15,12 @@ def calculate_hour_avg(df_data: pd.DataFrame) -> pd.DataFrame:
 
     by organwalk 2023-09-19
     """
+
     df_data['Time'] = pd.to_datetime(df_data['Time'])
     df = df_data.set_index('Time')
     df_hourly = df.resample('H').mean().round(2).reset_index().drop('Time', axis=1)
 
-    return df_hourly.astype(int)
+    return df_hourly
 
 
 def get_scaler_result(df_data: pd.DataFrame) -> (pd.DataFrame, MinMaxScaler):
